test(action-sheet): migrate a11y test
